move Panel to a different file
