+++ src/game/world/mod.rs
@@ -1,6 +1,6 @@
 use std::collections::HashMap;
 
-use graphics::renderer::Panel;
+use graphics::panel::Panel;
 use util::vector::Vector2D;
 
 mod chunk;
