+++ src/graphics/mod.rs
@@ -1,4 +1,5 @@
 #[macro_use]
 pub mod colour;
+pub mod panel;
 pub mod renderer;
 pub mod sprite;
